fix(ESF): #4525 Beautified condition name in custom dialog. (#4528)
* fix(ESF): #4525 Beautified condition name in custom dialog.

* fix(tests): #4503 Refactorred filtering tests.

* test(Filtering): #4503 Fixed tests after merge.

* test(Filter): #4525 Fixed lint warning.
